+++ itk_dev_shared_components/eflyt/eflyt_login.py
@@ -1,10 +1,25 @@
 """Module for logging into Eflyt/Daedalus/Whatchamacallit using Selenium"""
+import time
+
 from selenium import webdriver
 from selenium.webdriver.common.by import By
-from selenium.common.exceptions import NoSuchElementException
+from selenium.webdriver.remote.webelement import WebElement
+from selenium.common.exceptions import NoSuchElementException, StaleElementReferenceException
 
 
-def login(username: str, password: str) -> webdriver.Chrome:
+class ResilientBrowser(webdriver.Chrome):
+    """A webdriver.Chrome subclass that retries find_element if the element goes stale."""
+    def find_element(self, by=By.ID, value = None) -> WebElement:
+        element = super().find_element(by, value)
+        time.sleep(0.1)
+        try:
+            element.tag_name
+        except StaleElementReferenceException:
+            element = super().find_element(by, value)
+        return element
+
+
+def login(username: str, password: str) -> ResilientBrowser:
     """Log into Eflyt using a password and username.
 
     Args:
@@ -13,8 +28,9 @@ def login(username: str, password: str) -> webdriver.Chrome:
     """
     chrome_options = webdriver.ChromeOptions()
     chrome_options.add_argument("--disable-search-engine-choice-screen")
-    browser = webdriver.Chrome(options=chrome_options)
+    browser = ResilientBrowser(options=chrome_options)
     browser.maximize_window()
+    browser.implicitly_wait(2)
 
     browser.get("https://notuskommunal.scandihealth.net/")
     browser.find_element(By.ID, "Login1_UserName").send_keys(username)